+++ src/main/java/com/airtribe/learntrack/entity/Person.java
@@ -3,7 +3,7 @@ package com.airtribe.learntrack.entity;
 import com.airtribe.learntrack.util.InputValidator;
 
 public class Person {
-    private Long id;
+    private int id;
     private String firstName;
     private String lastName;
     private String email;
@@ -11,18 +11,18 @@ public class Person {
     public Person() {
     }
 
-    public Person(Long id, String firstName, String lastName, String email) {
+    public Person(int id, String firstName, String lastName, String email) {
         setId(id);
         setFirstName(firstName);
         setLastName(lastName);
         setEmail(email);
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         InputValidator.validateId(id, "ID");
         this.id = id;
     }
